Development Log Entry Date:18/11/2025 homework number: 1 1. Goal for today (C i) State what your plan said you should complete during this session. My goal was to get the form submission working and all the required fields set up. I also planned on making sure there was form validation with flash messages and then testing it to see if it worked. 2. Work completed (C iii) Write one or two sentences summarising what you achieved in this session. I finished adding all the fields that I needed for the form but its possible I will add more later on if something important seems to be missing. Before setting up the validation I set up the JSON saving. Then I set up the flash validation messages to check if the field is filled out and if it isn't then it displays a message above the field. I then tested the validation my leaving some fields empty and it worked.
3. Did you follow your plan? (C iii and C iv)
No:
I had to set up the JSON saving and the view page before the validation because the way I was validating was: if os.path.exists('registrations.json') so It was easier to just set up the JSON before. This saves time because if not I would have had to change the validation method later when I later wanted to implement the saving. This also helped me check if the validation worked because I could see the entries in the view page.

4. Next step (C i)
List one or two tasks you intend to complete during the next lesson.
I have already set up a view template more as a placeholder and to check if the submissions are actually being appended and next lesson I will work on improving it slightly to make it clearer.
that will be short so after I will probably start testing all features and fixing problems. If I find that I am missing an important field I will add it.